Extract ChangeBuffer module
