+++ lib/steep.rb
@@ -87,6 +87,7 @@ require "steep/index/signature_symbol_provider"
 require "steep/index/source_index"
 
 require "steep/server/utils"
+require "steep/server/change_buffer"
 require "steep/server/base_worker"
 require "steep/server/worker_process"
 require "steep/server/interaction_worker"
